pyproject.toml
[project]
name = "multi-agent-task-breakdown"
version = "0.1.0"
description = "Multi-Agent Task Breakdown with LangChain/LangGraph and AWS Bedrock"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-aws>=0.1.0",
    "langchain-community>=0.1.0",
    "langgraph>=0.1.0",
    "boto3>=1.34.0",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py311"
